Move FilterMatcher etc to its own submodule.
